+++ lib/cmdx/log_formatters/line.rb
@@ -5,8 +5,10 @@ module CMDx
     class Line
 
       def call(severity, time, progname, message)
+        time    = Utils::DatetimeFormatter.call(time.utc)
         message = message.to_h.map { |k, v| "#{k}=#{v}" }.join(" ") if message.is_a?(Result)
-        "#{Utils::ColorSeverity.call(severity[0])}, [#{Utils::DatetimeFormatter.call(time.utc)} ##{Process.pid}] #{Utils::ColorSeverity.call(severity)} -- #{progname || 'CMDx'}: #{message}\n"
+
+        "#{severity[0]}, [#{time} ##{Process.pid}] #{severity} -- #{progname || 'CMDx'}: #{message}\n"
       end
 
     end