+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## [Unreleased]
 
-- Add log severity coloring
+- Add PrettyLine log formatter with ansi color on severity, state, and status
 - Fix not being able to log non hash values
 
 ## [0.1.0] - 2025-03-07